Added the master makefile.
To build the library, do:
make -f master_makefile
To clean the library do:
make clean -f master_makefile

And so on, you get the idea.

The situation is as it is because this master makefile is intended
to be external to the library i.e. to manage the building of many
projects.

As a matter of fact, on my configuration I have a seperate master
makefile, and the one added now is a coppied and cropped version
of it.